[skip ci] fix typo (#722)
[skip ci] fix typo (#722)

+++ python/arctern/plot/_plot_with_map.py
@@ -55,7 +55,7 @@ def pointmap(ax, points, bounding_box,
     :param bounding_box: Specify the bounding rectangle [west, south, east, north].
 
     :type point_size: int
-    :param point_size: Diameter of point, default as 3 
+    :param point_size: Diameter of point, default as 3
 
     :type point_color: str
     :param point_color: Specify point color in Hex Color Code, default as "#115f9a"
@@ -124,7 +124,7 @@ def weighted_pointmap(ax, points, color_weights=None,
     :param points: Sequence of Points
 
     :type color_weights: Series(dtype: float|int64)
-    :param color_weights: Weights for point color, default as None 
+    :param color_weights: Weights for point color, default as None
 
     :type size_weights: Series(dtype: float|int64)
     :param size_weights: Weights for point size, deciding diameter of point (after bounded by size_bound)
@@ -135,21 +135,21 @@ def weighted_pointmap(ax, points, color_weights=None,
                          Default as None
 
     :type color_gradient: list
-    :param color_gradient: Specify range of color gradient. 
-                           Either use ["hex_color"] to specify a same color for all points, 
+    :param color_gradient: Specify range of color gradient.
+                           Either use ["hex_color"] to specify a same color for all points,
                            or ["hex_color1", "hex_color2"] to specify a color gradient ranging from "hex_color1" to "hex_color2"
                            Default as ["#115f9a", "#d0f400"]
-    
+
     :type color_bound: list
-    :param color_bound: Specify weight range [w1, w2] binding to color_gradient. 
-                        Needed only when color_gradient has two value ["color1", "color2"]. 
+    :param color_bound: Specify weight range [w1, w2] binding to color_gradient.
+                        Needed only when color_gradient has two value ["color1", "color2"].
                         Bind w1 to "color1", and w2 to "color2".
-                        When weight < w1 or weight > w2, truncate to w1/w2 accordingly. 
+                        When weight < w1 or weight > w2, truncate to w1/w2 accordingly.
                         Default as [0, 0]
 
     :type size_bound: list
-    :param size_bound: Specify range [w1, w2] of size_weights. 
-                       When weight < w1 or weight > w2, truncate to w1/w2 accordingly. 
+    :param size_bound: Specify range [w1, w2] of size_weights.
+                       When weight < w1 or weight > w2, truncate to w1/w2 accordingly.
                        Default as [3]
 
     :type opacity: float
@@ -173,17 +173,17 @@ def weighted_pointmap(ax, points, color_weights=None,
     >>> # Download link: https://raw.githubusercontent.com/arctern-io/arctern-resources/benchmarks/benchmarks/dataset/layer_rendering_test_data/test_data.csv
     >>> df = pd.read_csv("/path/to/test_data.csv", dtype={'longitude':np.float64, 'latitude':np.float64, 'color_weights':np.float64, 'size_weights':np.float64, 'region_boundaries':np.object})
     >>> points = arctern.GeoSeries.point(df['longitude'], df['latitude'])
-    >>> 
+    >>>
     >>> # plot weighted pointmap with variable color and fixed size
     >>> fig, ax = plt.subplots(figsize=(10, 6), dpi=200)
     >>> arctern.plot.weighted_pointmap(ax, points, color_weights=df['color_weights'], bounding_box=[-73.99668712186558,40.72972339069935,-73.99045479584949,40.7345193345495], color_gradient=["#115f9a", "#d0f400"], color_bound=[2.5,15], size_bound=[16], opacity=1.0, coordinate_system="EPSG:4326")
     >>> plt.show()
-    >>> 
+    >>>
     >>> # plot weighted pointmap with fixed color and variable size
     >>> fig, ax = plt.subplots(figsize=(10, 6), dpi=200)
     >>> arctern.plot.weighted_pointmap(ax, points, size_weights=df['size_weights'], bounding_box=[-73.99668712186558,40.72972339069935,-73.99045479584949,40.7345193345495], color_gradient=["#37A2DA"], size_bound=[15, 50], opacity=1.0, coordinate_system="EPSG:4326")
     >>> plt.show()
-    >>> 
+    >>>
     >>> # plot weighted pointmap with variable color and size
     >>> fig, ax = plt.subplots(figsize=(10, 6), dpi=200)
     >>> arctern.plot.weighted_pointmap(ax, points, color_weights=df['color_weights'], size_weights=df['size_weights'], bounding_box=[-73.99668712186558,40.72972339069935,-73.99045479584949,40.7345193345495], color_gradient=["#115f9a", "#d0f400"], color_bound=[2.5,15], size_bound=[15, 50], opacity=1.0, coordinate_system="EPSG:4326")
@@ -235,8 +235,8 @@ def heatmap(ax, points, weights, bounding_box,
     :type bounding_box: list
     :param bounding_box: Specify the bounding rectangle [west, south, east, north].
 
-    :type map_zoom_level: float
-    :param map_zoom_level: Zoom level of heatmap. Default as auto. 
+    :type map_zoom_level: int
+    :param map_zoom_level: Zoom level of heatmap. Default as auto.
 
     :type coordinate_system: str
     :param coordinate_system: Coordinate Reference System of the geometry objects.
@@ -259,7 +259,7 @@ def heatmap(ax, points, weights, bounding_box,
     >>> # Download link: https://raw.githubusercontent.com/arctern-io/arctern-resources/benchmarks/benchmarks/dataset/layer_rendering_test_data/test_data.csv
     >>> df = pd.read_csv("/path/to/test_data.csv", dtype={'longitude':np.float64, 'latitude':np.float64, 'color_weights':np.float64, 'size_weights':np.float64, 'region_boundaries':np.object})
     >>> points = arctern.GeoSeries.point(df['longitude'], df['latitude'])
-    >>> 
+    >>>
     >>> # plot heatmap
     >>> fig, ax = plt.subplots(figsize=(10, 6), dpi=200)
     >>> arctern.plot.heatmap(ax, points, df['color_weights'], bounding_box=[-74.01424568752932, 40.72759334104623, -73.96056823889673, 40.76721122683304], coordinate_system='EPSG:4326')
@@ -304,15 +304,15 @@ def choroplethmap(ax, region_boundaries, weights, bounding_box,
     :param bounding_box: Specify the bounding rectangle [west, south, east, north].
 
     :type color_gradient: list
-    :param color_gradient: Specify range of color gradient. 
-                           Either use ["hex_color"] to specify a same color for all polygons, 
+    :param color_gradient: Specify range of color gradient.
+                           Either use ["hex_color"] to specify a same color for all polygons,
                            or ["hex_color1", "hex_color2"] to specify a color gradient ranging from "hex_color1" to "hex_color2"
-    
+
     :type color_bound: list
-    :param color_bound: Specify weight range [w1, w2] binding to color_gradient. 
-                        Needed only when color_gradient has two value ["color1", "color2"]. 
+    :param color_bound: Specify weight range [w1, w2] binding to color_gradient.
+                        Needed only when color_gradient has two value ["color1", "color2"].
                         Bind w1 to "color1", and w2 to "color2".
-                        When weight < w1 or weight > w2, truncate to w1/w2 accordingly. 
+                        When weight < w1 or weight > w2, truncate to w1/w2 accordingly.
                         Default as None
 
     :type opacity: float
@@ -444,8 +444,8 @@ def fishnetmap(ax, points, weights, bounding_box,
     :param bounding_box: Specify the bounding rectangle [west, south, east, north].
 
     :type color_gradient: list
-    :param color_gradient: Specify range of color gradient. 
-                           Either use ["hex_color"] to specify a same color for all points, 
+    :param color_gradient: Specify range of color gradient.
+                           Either use ["hex_color"] to specify a same color for all points,
                            or ["hex_color1", "hex_color2"] to specify a color gradient ranging from "hex_color1" to "hex_color2"
                            Current only default value ["#0000FF", "#FF0000"] is supported
